refactor: extract calendar layouts to layouts/calendar.py
